pyproject.toml
[tool.poetry]
name = "PairCodeTests"
version = "0.1.0"
description = ""
authors = ["ThaRising <mail@domain.com>"]

[tool.poetry.dependencies]
python = "^3.8"
django = "^3.0.8"
djangorestframework = "^3.11.0"
drf-yasg = "^1.17.1"
django-crispy-forms = "^1.9.2"
Pillow = "^7.2.0"
django-debug-toolbar = "^2.2"

[tool.poetry.dev-dependencies]
isort = "^5.1.4"
autoflake = "^1.3.1"
black = "^19.10b0"

[tool.black]
line-length = 80
target_version = ["py38"]
include = '\.pyi?$'

[tool.isort]
known_first_party = ["PairCodeTests", "tests"]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
combine_as_imports = true

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
